DartGame
This code simulates a game of darts by asking the user for the number of throws and generating random coordinates for each dart. It counts how many darts land in zones A, B, C or D and ends if one dart misses all zones. Finally it displays the number and percentage of darts in each zone.